LPD-94529 Consolidate PageSpeed components in the pagespeed folder
