+++ modules/dxp/apps/seo-studio/seo-studio-web/src/main/resources/META-INF/resources/js/index.ts
@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default as PageSpeedCharts} from './PageSpeedCharts';
 export {default as SectionHeader} from './components/SectionHeader';
 export {default as HealthScanConfiguration} from './health_scan/components/HealthScanConfiguration';
 export {default as InsightDetailView} from './insights_view/InsightDetailView';
 export {default as OnPage} from './on_page/OnPage';
+export {default as PageSpeedCharts} from './pagespeed/PageSpeedCharts';
 export {default as PageSpeedConfiguration} from './pagespeed/PageSpeedConfiguration';
