+++ modules/dxp/apps/seo-studio/seo-studio-web/src/main/resources/META-INF/resources/js/pagespeed/PageSpeedCharts.tsx
