+++ modules/dxp/apps/seo-studio/seo-studio-web/src/main/resources/META-INF/resources/tests/js/pagespeed/GaugeChart.spec.tsx
@@ -7,7 +7,7 @@ import '@testing-library/jest-dom';
 import {render, screen} from '@testing-library/react';
 import React from 'react';
 
-import GaugeChart from '../../../js/GaugeChart';
+import GaugeChart from '../../../js/pagespeed/GaugeChart';
 
 describe('GaugeChart', () => {
 	it('renders the label passed in', () => {
